Widgets: take the toolkit half of gtk.c, and cut its last globals
gtk.c was ~3400 lines, of which about half had nothing to do with the application: container
child counting, label construction, popovers, notebooks, collapsible sections, resizable
scroll wrappers, startup dialogs. They lived there because gtk.c is where GUI code went, and
consequently 140 files included gui/gtk.h to reach them.

They move to widgets/, one module per topic rather than one more god-header:

  container.c            counting and clearing a GtkContainer's children
  label.c                the two standard label shapes, markup stripping, symbolic icons
  popup.c                menus, popovers, and the Wayland top-most-parent rule
  notebook.c             tabs that size to the available width
  collapsible_section.c  a section that folds away and remembers it
  scroll_wrap.c          drag-to-resize scrolled windows and drawing areas
  dialog.c               the three standalone dialogs that run before the GUI exists

and the remainder joins the module that already owns its subject: search-entry handlers to
gtkentry, the accels-during-text-input guard to accelerators, the DPI/modifier probes and
the freeze macros to widget_settings.

The point was not relocation but decoupling, so each moved function lost its reach into the
application. Four of them read darktable.gui->ui->main_window to find a window to anchor to;
dt_widget_root_window() already answers that and already returns NULL when there is none, so
a three-deep global guard collapsed into the NULL check that followed it. Three more needed
genuinely new seams, all following the handler pattern already in widget_settings: a refocus
request (Escape in a text entry hands focus back to whatever the host considers its main
area), a notebook page-change report (the host relays it on its signal bus; widgets/ has no
bus), and per-widget persistence (a panel remembers the height you dragged it to, under a key
you supplied -- where that is stored is the host's business).

Two globals were reached implicitly rather than by name, and both are now gone from widgets/:

  common/conf.h  a configuration system is application state behind a global, whichever
                 function you call to read it. Replaced by the persistence handlers above:
                 unregistered, nothing is stored and widgets fall back to their defaults,
                 which is also correct for anything embedding these widgets standalone.

  common/logging.h  dt_print() means a global flags word and a global stream. widgets/ now
                 prints straight to stdout in a debug build and compiles the call away
                 otherwise. Nothing is lost: these messages describe widget internals, and
                 whoever reads them is running a debug build anyway.

Result: nothing under widgets/ includes gui/, develop/, control/, libs/ or views/, and
nothing there names an application global. Verified by grep and by compiling the _DEBUG paths
separately, since the release build never sees them.

gtk.h re-exports the new headers for now so the 140 existing includers keep building; moving
them onto the specific headers is the next step, and the reason the split was worth doing.

Both configurations build clean.
