+++ src/gui/gtk.c
@@ -140,37 +140,6 @@ static void _ui_toast_redraw_callback(gpointer instance, GtkWidget *widget);
 // set class function to add CSS classes with just a simple line call
 
 
-void dt_gui_set_symbolic_icon(GtkWidget *image, const char *icon_name, GtkIconSize size, const GdkRGBA *color)
-{
-  gint width = 16, height = 16;
-  gtk_icon_size_lookup(size, &width, &height);
-  // Icon themes only look up square icons: request the larger dimension,
-  // then scale the result down to the exact width/height below if the two differ.
-  GtkIconInfo *info = gtk_icon_theme_lookup_icon(gtk_icon_theme_get_default(), icon_name, MAX(width, height),
-                                                 GTK_ICON_LOOKUP_FORCE_SYMBOLIC);
-  if(IS_NULL_PTR(info))
-  {
-    gtk_image_set_from_icon_name(GTK_IMAGE(image), icon_name, size);
-    return;
-  }
-
-  GdkPixbuf *pixbuf = IS_NULL_PTR(color)
-      ? gtk_icon_info_load_symbolic_for_context(info, gtk_widget_get_style_context(image), NULL, NULL)
-      : gtk_icon_info_load_symbolic(info, color, color, color, color, NULL, NULL);
-  g_object_unref(info);
-
-  if(!IS_NULL_PTR(pixbuf))
-  {
-    if(gdk_pixbuf_get_width(pixbuf) != width || gdk_pixbuf_get_height(pixbuf) != height)
-    {
-      GdkPixbuf *scaled = gdk_pixbuf_scale_simple(pixbuf, width, height, GDK_INTERP_BILINEAR);
-      g_object_unref(pixbuf);
-      pixbuf = scaled;
-    }
-    gtk_image_set_from_pixbuf(GTK_IMAGE(image), pixbuf);
-    g_object_unref(pixbuf);
-  }
-}
 
 /* ------------------------------------------------------------------------------------------
  * Widget-callback suppression depth (see darktable.h for the rationale and API).
@@ -965,6 +934,34 @@ static gint _widget_natural_width(GtkWidget *widget)
 
 /* dt_control_change_cursor() and dt_toast_log() are macros, so the widget hooks need real
  * functions to point at. */
+// The host owns the preferences; widgets only know the key they were given.
+static gboolean _widget_stored_int(const char *key, int *value)
+{
+  if(!dt_conf_key_exists(key)) return FALSE;
+  *value = dt_conf_get_int(key);
+  return TRUE;
+}
+
+static void _widget_store_int(const char *key, int value)
+{
+  dt_conf_set_int(key, value);
+}
+
+static gboolean _widget_stored_bool(const char *key)
+{
+  return dt_conf_get_bool(key);
+}
+
+static void _widget_store_bool(const char *key, gboolean value)
+{
+  dt_conf_set_bool(key, value);
+}
+
+static void _notebook_page_changed(gpointer owner)
+{
+  DT_DEBUG_CONTROL_SIGNAL_RAISE(dt_control_signal_get_global(), DT_SIGNAL_CONTROL_NOTEBOOK_TAB_CHANGED, owner);
+}
+
 static void _widget_cursor(GdkCursorType cursor)
 {
   dt_control_change_cursor(cursor);
@@ -1217,6 +1214,11 @@ int dt_gui_gtk_init(dt_gui_gtk_t *gui)
   dt_widget_set_root_window_handler(_widget_root_window);
   dt_widget_set_natural_width_handler(_widget_natural_width);
   dt_widget_set_cursor_handler(_widget_cursor);
+  dt_widget_set_refocus_handler(dt_gui_refocus_center);
+  dt_widget_set_debug_overlays((dt_get_debug_flags() & DT_DEBUG_MASKS) != 0);
+  dt_widget_set_storage_handlers(_widget_stored_int, _widget_store_int,
+                                 _widget_stored_bool, _widget_store_bool);
+  dt_widget_set_notebook_page_handler(_notebook_page_changed);
   dt_widget_set_message_handler(_widget_message);
   dt_accels_set_global(gui->accels);
   dt_accels_set_top_offset_handler(_accels_top_offset);
@@ -1272,24 +1274,6 @@ void dt_gui_gtk_run(dt_gui_gtk_t *gui)
   dt_cleanup();
 }
 
-// refactored function to read current ppd, because gtk for osx has been unreliable
-// we use the specific function here. Anyway, if nothing meaningful is found we default back to 1.0
-double dt_get_system_gui_ppd(GtkWidget *widget)
-{
-  double res = 0.0f;
-#ifdef GDK_WINDOWING_QUARTZ
-  res = dt_osx_get_ppd();
-#else
-  res = gtk_widget_get_scale_factor(widget);
-#endif
-  if((res < 1.0f) || (res > 4.0f))
-  {
-    dt_print(DT_DEBUG_CONTROL, "[dt_get_system_gui_ppd] can't detect system ppd\n");
-    return 1.0f;
-  }
-  dt_print(DT_DEBUG_CONTROL, "[dt_get_system_gui_ppd] system ppd is %f\n", res);
-  return res;
-}
 
 void dt_configure_ppd_dpi(dt_gui_gtk_t *gui)
 {
@@ -1735,341 +1719,15 @@ static void _ui_toast_redraw_callback(gpointer instance, GtkWidget *widget)
   dt_pthread_mutex_unlock(&control->toast_mutex);
 }
 
-void dt_ellipsize_combo(GtkComboBox *cbox)
-{
-  GList *renderers = gtk_cell_layout_get_cells(GTK_CELL_LAYOUT(cbox));
-  for(const GList *it = renderers; it; it = g_list_next(it))
-  {
-    GtkCellRendererText *tr = GTK_CELL_RENDERER_TEXT(it->data);
-    g_object_set(G_OBJECT(tr), "ellipsize", PANGO_ELLIPSIZE_MIDDLE, (gchar *)0);
-  }
-  g_list_free(renderers);
-  renderers = NULL;
-}
 
-typedef struct result_t
-{
-  enum {RESULT_NONE, RESULT_NO, RESULT_YES} result;
-  char *entry_text;
-  GtkWidget *window, *entry, *button_yes, *button_no;
-} result_t;
 
-static void _gtk_main_quit_safe(GtkWidget *widget, gpointer data)
-{
-  (void)widget;
-  (void)data;
-  if(gtk_main_level() > 0) gtk_main_quit();
-}
 
-static void _yes_no_button_handler(GtkButton *button, gpointer data)
-{
-  result_t *result = (result_t *)data;
 
-  if((void *)button == (void *)result->button_yes)
-    result->result = RESULT_YES;
-  else if((void *)button == (void *)result->button_no)
-    result->result = RESULT_NO;
 
-  if(result->entry)
-    result->entry_text = g_strdup(gtk_entry_get_text(GTK_ENTRY(result->entry)));
-  gtk_widget_destroy(result->window);
-  _gtk_main_quit_safe(NULL, NULL);
-}
 
-void dt_gui_refocus_parent(GtkWindow *parent)
-{
-  if(!GTK_IS_WINDOW(parent) && !IS_NULL_PTR(darktable.gui) && !IS_NULL_PTR(darktable.gui->ui)
-     && !IS_NULL_PTR(darktable.gui->ui->main_window))
-  {
-    GtkWidget *main_window = dt_ui_main_window(darktable.gui->ui);
-    if(GTK_IS_WINDOW(main_window)) parent = GTK_WINDOW(main_window);
-  }
 
-  if(GTK_IS_WINDOW(parent)) gtk_window_present(parent);
 
-#ifdef GDK_WINDOWING_QUARTZ
-  dt_osx_focus_window();
-#endif
-}
 
-gboolean dt_gui_show_standalone_yes_no_dialog(const char *title, const char *markup, const char *no_text,
-                                              const char *yes_text)
-{
-  GtkWidget *window = gtk_window_new(GTK_WINDOW_TOPLEVEL);
-#ifdef GDK_WINDOWING_QUARTZ
-  dt_osx_disallow_fullscreen(window);
-#endif
-
-  // themes not yet loaded, no CSS add some manual padding
-  const int padding = dt_gui_get_themes() ? 0 : 5;
-
-  gtk_window_set_icon_name(GTK_WINDOW(window), "ansel");
-  gtk_window_set_title(GTK_WINDOW(window), title);
-  g_signal_connect(window, "destroy", G_CALLBACK(_gtk_main_quit_safe), NULL);
-
-  gtk_window_set_position(GTK_WINDOW(window), GTK_WIN_POS_CENTER);
-
-  if(!IS_NULL_PTR(darktable.gui) && !IS_NULL_PTR(darktable.gui->ui)
-     && !IS_NULL_PTR(darktable.gui->ui->main_window))
-  {
-    GtkWidget *main_window = dt_ui_main_window(darktable.gui->ui);
-    if(GTK_IS_WINDOW(main_window))
-    {
-      GtkWindow *win = GTK_WINDOW(main_window);
-      gtk_window_set_transient_for(GTK_WINDOW(window), win);
-      gtk_window_set_modal(GTK_WINDOW(window), TRUE);
-      if(gtk_widget_get_visible(GTK_WIDGET(win)))
-      {
-        gtk_window_set_position(GTK_WINDOW(window), GTK_WIN_POS_CENTER_ON_PARENT);
-      }
-    }
-  }
-
-  GtkWidget *vbox = gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING);
-  gtk_container_add(GTK_CONTAINER(window), vbox);
-
-  GtkWidget *mhbox = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, DT_GUI_BOX_SPACING);
-  gtk_box_pack_start(GTK_BOX(vbox), mhbox, TRUE, TRUE, padding);
-
-  if(padding)
-  {
-    gtk_box_pack_start(GTK_BOX(mhbox),
-                       gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING), TRUE, TRUE, padding);
-  }
-
-  GtkWidget *label = gtk_label_new(NULL);
-  gtk_label_set_markup(GTK_LABEL(label), markup);
-  gtk_box_pack_start(GTK_BOX(mhbox), label, TRUE, TRUE, padding);
-
-  if(padding)
-  {
-    gtk_box_pack_start(GTK_BOX(mhbox),
-                       gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING), TRUE, TRUE, padding);
-  }
-
-  GtkWidget *hbox = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, DT_GUI_BOX_SPACING);
-  gtk_box_pack_start(GTK_BOX(vbox), hbox, TRUE, TRUE, 0);
-
-  result_t result = {.result = RESULT_NONE, .window = window};
-
-  GtkWidget *button;
-
-  if(no_text)
-  {
-    button = gtk_button_new_with_label(no_text);
-    result.button_no = button;
-    g_signal_connect(G_OBJECT(button), "clicked", G_CALLBACK(_yes_no_button_handler), &result);
-    gtk_box_pack_start(GTK_BOX(hbox), button, TRUE, TRUE, 0);
-  }
-
-  if(yes_text)
-  {
-    button = gtk_button_new_with_label(yes_text);
-    result.button_yes = button;
-    g_signal_connect(G_OBJECT(button), "clicked", G_CALLBACK(_yes_no_button_handler), &result);
-    gtk_box_pack_start(GTK_BOX(hbox), button, TRUE, TRUE, 0);
-  }
-
-  gtk_widget_show_all(window);
-  gtk_main();
-
-  return result.result == RESULT_YES;
-}
-
-typedef struct _three_choice_result_t
-{
-  int result;
-  GtkWidget *window, *button_first, *button_second, *button_third;
-} _three_choice_result_t;
-
-static void _three_choice_button_handler(GtkButton *button, gpointer data)
-{
-  _three_choice_result_t *result = (_three_choice_result_t *)data;
-
-  if((void *)button == (void *)result->button_first)
-    result->result = 0;
-  else if((void *)button == (void *)result->button_second)
-    result->result = 1;
-  else if((void *)button == (void *)result->button_third)
-    result->result = 2;
-
-  gtk_widget_destroy(result->window);
-  _gtk_main_quit_safe(NULL, NULL);
-}
-
-int dt_gui_show_standalone_three_choice_dialog(const char *title, const char *markup, const char *first_text,
-                                               const char *second_text, const char *third_text)
-{
-  GtkWidget *window = gtk_window_new(GTK_WINDOW_TOPLEVEL);
-#ifdef GDK_WINDOWING_QUARTZ
-  dt_osx_disallow_fullscreen(window);
-#endif
-
-  // themes not yet loaded, no CSS add some manual padding
-  const int padding = dt_gui_get_themes() ? 0 : 5;
-
-  gtk_window_set_icon_name(GTK_WINDOW(window), "ansel");
-  gtk_window_set_title(GTK_WINDOW(window), title);
-  g_signal_connect(window, "destroy", G_CALLBACK(_gtk_main_quit_safe), NULL);
-
-  gtk_window_set_position(GTK_WINDOW(window), GTK_WIN_POS_CENTER);
-
-  if(!IS_NULL_PTR(darktable.gui) && !IS_NULL_PTR(darktable.gui->ui)
-     && !IS_NULL_PTR(darktable.gui->ui->main_window))
-  {
-    GtkWidget *main_window = dt_ui_main_window(darktable.gui->ui);
-    if(GTK_IS_WINDOW(main_window))
-    {
-      GtkWindow *win = GTK_WINDOW(main_window);
-      gtk_window_set_transient_for(GTK_WINDOW(window), win);
-      gtk_window_set_modal(GTK_WINDOW(window), TRUE);
-      if(gtk_widget_get_visible(GTK_WIDGET(win)))
-      {
-        gtk_window_set_position(GTK_WINDOW(window), GTK_WIN_POS_CENTER_ON_PARENT);
-      }
-    }
-  }
-
-  GtkWidget *vbox = gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING);
-  gtk_container_add(GTK_CONTAINER(window), vbox);
-
-  GtkWidget *mhbox = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, DT_GUI_BOX_SPACING);
-  gtk_box_pack_start(GTK_BOX(vbox), mhbox, TRUE, TRUE, padding);
-
-  if(padding)
-  {
-    gtk_box_pack_start(GTK_BOX(mhbox),
-                       gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING), TRUE, TRUE, padding);
-  }
-
-  GtkWidget *label = gtk_label_new(NULL);
-  gtk_label_set_markup(GTK_LABEL(label), markup);
-  gtk_box_pack_start(GTK_BOX(mhbox), label, TRUE, TRUE, padding);
-
-  if(padding)
-  {
-    gtk_box_pack_start(GTK_BOX(mhbox),
-                       gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING), TRUE, TRUE, padding);
-  }
-
-  GtkWidget *hbox = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, DT_GUI_BOX_SPACING);
-  gtk_box_pack_start(GTK_BOX(vbox), hbox, TRUE, TRUE, 0);
-
-  _three_choice_result_t result = { .result = -1, .window = window };
-
-  GtkWidget *button;
-
-  if(first_text)
-  {
-    button = gtk_button_new_with_label(first_text);
-    result.button_first = button;
-    g_signal_connect(G_OBJECT(button), "clicked", G_CALLBACK(_three_choice_button_handler), &result);
-    gtk_box_pack_start(GTK_BOX(hbox), button, TRUE, TRUE, 0);
-  }
-
-  if(second_text)
-  {
-    button = gtk_button_new_with_label(second_text);
-    result.button_second = button;
-    g_signal_connect(G_OBJECT(button), "clicked", G_CALLBACK(_three_choice_button_handler), &result);
-    gtk_box_pack_start(GTK_BOX(hbox), button, TRUE, TRUE, 0);
-  }
-
-  if(third_text)
-  {
-    button = gtk_button_new_with_label(third_text);
-    result.button_third = button;
-    g_signal_connect(G_OBJECT(button), "clicked", G_CALLBACK(_three_choice_button_handler), &result);
-    gtk_box_pack_start(GTK_BOX(hbox), button, TRUE, TRUE, 0);
-  }
-
-  gtk_widget_show_all(window);
-  gtk_main();
-
-  return result.result;
-}
-
-char *dt_gui_show_standalone_string_dialog(const char *title, const char *markup, const char *placeholder,
-                                           const char *no_text, const char *yes_text)
-{
-  GtkWidget *window = gtk_window_new(GTK_WINDOW_TOPLEVEL);
-#ifdef GDK_WINDOWING_QUARTZ
-  dt_osx_disallow_fullscreen(window);
-#endif
-
-  gtk_window_set_icon_name(GTK_WINDOW(window), "ansel");
-  gtk_window_set_title(GTK_WINDOW(window), title);
-  g_signal_connect(window, "destroy", G_CALLBACK(_gtk_main_quit_safe), NULL);
-
-  gtk_window_set_position(GTK_WINDOW(window), GTK_WIN_POS_MOUSE);
-
-  if(!IS_NULL_PTR(darktable.gui) && !IS_NULL_PTR(darktable.gui->ui)
-     && !IS_NULL_PTR(darktable.gui->ui->main_window))
-  {
-    GtkWidget *main_window = dt_ui_main_window(darktable.gui->ui);
-    if(GTK_IS_WINDOW(main_window))
-    {
-      GtkWindow *win = GTK_WINDOW(main_window);
-      gtk_window_set_transient_for(GTK_WINDOW(window), win);
-      if(gtk_widget_get_visible(GTK_WIDGET(win)))
-      {
-        gtk_window_set_position(GTK_WINDOW(window), GTK_WIN_POS_CENTER_ON_PARENT);
-      }
-    }
-  }
-
-  GtkWidget *vbox = gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING);
-  gtk_widget_set_margin_start(vbox, 10);
-  gtk_widget_set_margin_end(vbox, 10);
-  gtk_widget_set_margin_top(vbox, 7);
-  gtk_widget_set_margin_bottom(vbox, 5);
-  gtk_container_add(GTK_CONTAINER(window), vbox);
-
-  GtkWidget *label = gtk_label_new(NULL);
-  gtk_label_set_markup(GTK_LABEL(label), markup);
-  gtk_box_pack_start(GTK_BOX(vbox), label, TRUE, TRUE, 0);
-
-  GtkWidget *entry = gtk_entry_new();
-  dt_accels_disconnect_on_text_input(entry);
-
-  g_object_ref(entry);
-  if(placeholder)
-    gtk_entry_set_placeholder_text(GTK_ENTRY(entry), placeholder);
-  gtk_box_pack_start(GTK_BOX(vbox), entry, TRUE, TRUE, 0);
-
-  GtkWidget *hbox = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, DT_GUI_BOX_SPACING);
-  gtk_widget_set_margin_top(hbox, 10);
-  gtk_box_pack_start(GTK_BOX(vbox), hbox, TRUE, TRUE, 0);
-
-  result_t result = {.result = RESULT_NONE, .window = window, .entry = entry};
-
-  GtkWidget *button;
-
-  if(no_text)
-  {
-    button = gtk_button_new_with_label(no_text);
-    result.button_no = button;
-    g_signal_connect(G_OBJECT(button), "clicked", G_CALLBACK(_yes_no_button_handler), &result);
-    gtk_box_pack_start(GTK_BOX(hbox), button, TRUE, TRUE, 0);
-  }
-
-  if(yes_text)
-  {
-    button = gtk_button_new_with_label(yes_text);
-    result.button_yes = button;
-    g_signal_connect(G_OBJECT(button), "clicked", G_CALLBACK(_yes_no_button_handler), &result);
-    gtk_box_pack_start(GTK_BOX(hbox), button, TRUE, TRUE, 0);
-  }
-
-  gtk_widget_show_all(window);
-  gtk_main();
-
-  if(result.result == RESULT_YES)
-    return result.entry_text;
-
-  dt_free(result.entry_text);
-  return NULL;
-}
 
 // TODO: should that go to another place than gtk.c?
 void dt_gui_add_help_link(GtkWidget *widget, char *link)
@@ -2232,1054 +1890,82 @@ void dt_gui_load_theme(const char *theme)
     }
   }
 
+  dt_widget_set_theme_loaded(TRUE);
+
   // The active theme/font may change the root font size, so refresh the cached em
   // that drives DT_GUI_BOX_SPACING.
   dt_gui_update_em();
 }
 
-GdkModifierType dt_key_modifier_state()
-{
-  guint state = 0;
-  GdkWindow *window = gtk_widget_get_window(dt_ui_main_window(darktable.gui->ui));
-  gdk_device_get_state(gdk_seat_get_pointer(gdk_display_get_default_seat(gdk_window_get_display(window))), window, NULL, &state);
-  return state;
-
-/* FIXME double check correct way of doing this (merge conflict with Input System NG 20210319)
-  GdkKeymap *keymap = gdk_keymap_get_for_display(gdk_display_get_default());
-  return gdk_keymap_get_modifier_state(keymap) & gdk_keymap_get_modifier_mask(keymap, GDK_MODIFIER_INTENT_DEFAULT_MOD_MASK);
-*/
-}
-
-static void _notebook_size_callback(GtkNotebook *notebook, GdkRectangle *allocation, gpointer *data)
-{
-  const int n = gtk_notebook_get_n_pages(notebook);
-  g_return_if_fail(n > 0);
-
-  GtkRequestedSize *sizes = g_malloc_n(n, sizeof(GtkRequestedSize));
-
-  for(int i = 0; i < n; i++)
-  {
-    sizes[i].data = gtk_notebook_get_tab_label(notebook, gtk_notebook_get_nth_page(notebook, i));
-    sizes[i].minimum_size = 0;
-    GtkRequisition natural_size;
-    gtk_widget_get_preferred_size(sizes[i].data, NULL, &natural_size);
-    sizes[i].natural_size = natural_size.width;
-  }
-
-  GtkAllocation first, last;
-  gtk_widget_get_allocation(sizes[0].data, &first);
-  gtk_widget_get_allocation(sizes[n - 1].data, &last);
-
-  const gint total_space = last.x + last.width - first.x; // ignore tab padding; CSS sets padding for label
-
-  if(total_space > 0)
-  {
-    gtk_distribute_natural_allocation(total_space, n, sizes);
-
-    for(int i = 0; i < n; i++)
-      gtk_widget_set_size_request(sizes[i].data, sizes[i].minimum_size, -1);
-
-    gtk_widget_size_allocate(GTK_WIDGET(notebook), allocation);
-
-    for(int i = 0; i < n; i++)
-      gtk_widget_set_size_request(sizes[i].data, -1, -1);
-  }
-
-  dt_free(sizes);
-}
-
-// GTK_STATE_FLAG_PRELIGHT does not seem to get set on the label on hover so
-// state-flags-changed cannot update dt_control_get_global()->element for shortcut mapping
-static gboolean _notebook_motion_notify_callback(GtkWidget *widget, GdkEventMotion *event, gpointer user_data)
-{
-  GtkAllocation notebook_alloc, label_alloc;
-  gtk_widget_get_allocation(widget, &notebook_alloc);
-
-  GtkNotebook *notebook = GTK_NOTEBOOK(widget);
-  const int n = gtk_notebook_get_n_pages(notebook);
-  for(int i = 0; i < n; i++)
-  {
-    gtk_widget_get_allocation(gtk_notebook_get_tab_label(notebook, gtk_notebook_get_nth_page(notebook, i)), &label_alloc);
-  }
-
-  return FALSE;
-}
-
-GtkNotebook *dt_ui_notebook_new()
-{
-  return GTK_NOTEBOOK(gtk_notebook_new());
-}
-
-GtkWidget *dt_ui_notebook_page(GtkNotebook *notebook, const char *text, const char *tooltip)
-{
-  gchar *text_cpy = g_strdup(_(text));
-  dt_capitalize_label(text_cpy);
-  GtkWidget *label = gtk_label_new(text_cpy);
-  dt_free(text_cpy);
-  GtkWidget *page = gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING);
-  if(strlen(text) > 2)
-    gtk_label_set_ellipsize(GTK_LABEL(label), PANGO_ELLIPSIZE_END);
-  gtk_widget_set_tooltip_text(label, tooltip ? tooltip : _(text));
-  gtk_widget_set_has_tooltip(GTK_WIDGET(notebook), FALSE);
-
-  gint page_num = gtk_notebook_append_page(notebook, page, label);
-  gtk_container_child_set(GTK_CONTAINER(notebook), page, "tab-expand", TRUE, "tab-fill", TRUE, NULL);
-  if(page_num == 1 &&
-     !g_signal_handler_find(G_OBJECT(notebook), G_SIGNAL_MATCH_FUNC, 0, 0, NULL, _notebook_size_callback, NULL))
-  {
-    g_signal_connect(G_OBJECT(notebook), "size-allocate", G_CALLBACK(_notebook_size_callback), NULL);
-    g_signal_connect(G_OBJECT(notebook), "motion-notify-event", G_CALLBACK(_notebook_motion_notify_callback), NULL);
-  }
-
-  return page;
-}
-
-static void _notebook_switch_page_signal_relay(GtkNotebook *notebook, GtkWidget *page, guint page_num,
-                                               gpointer user_data)
-{
-  DT_DEBUG_CONTROL_SIGNAL_RAISE(dt_control_signal_get_global(), DT_SIGNAL_CONTROL_NOTEBOOK_TAB_CHANGED, user_data);
-}
-
-void dt_ui_notebook_set_picker_owner(GtkNotebook *notebook, gpointer owner)
-{
-  if(IS_NULL_PTR(notebook)) return;
-
-  // Guard against connecting twice on the same notebook (e.g. a caller re-entering
-  // gui_init after a GUI rebuild); user_data is set per-connection, so re-registering
-  // a different owner on an already-relayed notebook would otherwise leak the old
-  // connection instead of updating it.
-  const gulong existing = g_signal_handler_find(G_OBJECT(notebook), G_SIGNAL_MATCH_FUNC, 0, 0, NULL,
-                                                G_CALLBACK(_notebook_switch_page_signal_relay), NULL);
-  if(existing) g_signal_handler_disconnect(G_OBJECT(notebook), existing);
-
-  g_signal_connect(G_OBJECT(notebook), "switch_page", G_CALLBACK(_notebook_switch_page_signal_relay), owner);
-}
-
-static gint _get_container_row_heigth(GtkWidget *w)
-{
-  gint height = DT_PIXEL_APPLY_DPI(10);
-
-  if(GTK_IS_TREE_VIEW(w))
-  {
-    gint row_height = 0;
-
-    const gint num_columns = gtk_tree_view_get_n_columns(GTK_TREE_VIEW(w));
-    for(int c = 0; c < num_columns; c++)
-    {
-      gint cell_height = 0;
-      gtk_tree_view_column_cell_get_size(gtk_tree_view_get_column(GTK_TREE_VIEW(w), c),
-                                        NULL, NULL, NULL, NULL, &cell_height);
-      if(cell_height > row_height) row_height = cell_height;
-    }
-    GValue separation = { G_TYPE_INT };
-    gtk_widget_style_get_property(w, "vertical-separator", &separation);
-
-    if(row_height > 0) height = row_height + g_value_get_int(&separation);
-  }
-  else if(GTK_IS_TEXT_VIEW(w))
-  {
-    PangoLayout *layout = gtk_widget_create_pango_layout(w, "X");
-    pango_layout_get_pixel_size(layout, NULL, &height);
-    g_object_unref(layout);
-  }
-  else
-  {
-    GtkWidget *child = dt_gui_container_first_child(GTK_CONTAINER(w));
-    if(child)
-    {
-      height = gtk_widget_get_allocated_height(child);
-    }
-  }
-
-  return height;
-}
-
-static const char *const DT_GUI_WIDGET_AUTO_HEIGHT_KEY = "dt-gui-widget-auto-height";
-
-// find the scrolled window parent of a treeview, if any
-static GtkWidget *_search_parent_scrolled_window(GtkWidget *w)
-{
-  if(!GTK_IS_WIDGET(w)) return NULL;
-  
-  GtkWidget *parent = w;
-  while(parent)
-  {
-    if(GTK_IS_SCROLLED_WINDOW(parent)) break;
-    parent = gtk_widget_get_parent(parent);
-  }
-
-  return GTK_IS_SCROLLED_WINDOW(parent) ? parent : NULL;
-}
-
-// Counts only visible items (those whose parents are expanded)
-static int _treeview_count_visible_rows(GtkTreeView *treeview, GtkTreeModel *model, GtkTreeIter *parent)
-{
-  if(!GTK_IS_TREE_MODEL(model)) return 0;
-
-  GtkTreeIter iter;
-  gboolean valid = parent ? gtk_tree_model_iter_children(model, &iter, parent)
-                          : gtk_tree_model_get_iter_first(model, &iter);
-  int count = 0;
-
-  while(valid)
-  {
-    count++;
-    
-    // If this item is expanded, recursively count its visible children
-    if(gtk_tree_model_iter_has_child(model, &iter))
-    {
-      GtkTreePath *path = gtk_tree_model_get_path(model, &iter);
-      if(path)
-      {
-        if(gtk_tree_view_row_expanded(treeview, path))
-        {
-          count += _treeview_count_visible_rows(treeview, model, &iter);
-        }
-        gtk_tree_path_free(path);
-      }
-    }
-    
-    valid = gtk_tree_model_iter_next(model, &iter);
-  }
-
-  return count;
-}
-
-static int _textview_count_visible_rows(GtkWidget *textview)
-{
-  if(!GTK_IS_TEXT_VIEW(textview)) return 0;
-
-  GtkTextBuffer *buffer = gtk_text_view_get_buffer(GTK_TEXT_VIEW(textview));
-  if(!GTK_IS_TEXT_BUFFER(buffer)) return 0;
-
-  // For text views, use the number of logical lines in the buffer.
-  return MAX(0, gtk_text_buffer_get_line_count(buffer));
-}
-
-static void _widget_auto_disconnect_model(dt_gui_widget_auto_height_t *state, GtkWidget *treeview)
-{
-  if(IS_NULL_PTR(state)) return;
-
-  GtkTreeModel *model = state->model;
-  if(model)
-  {
-    if(state->model_row_inserted)   g_signal_handler_disconnect(model, state->model_row_inserted);
-    if(state->model_row_deleted)    g_signal_handler_disconnect(model, state->model_row_deleted);
-    if(state->model_row_changed)    g_signal_handler_disconnect(model, state->model_row_changed);
-    if(state->model_rows_reordered) g_signal_handler_disconnect(model, state->model_rows_reordered);
-    g_object_remove_weak_pointer(G_OBJECT(model), (gpointer *)&state->model);
-  }
-
-  if(GTK_IS_TREE_VIEW(treeview))
-  {
-    if(state->model_row_expanded)   g_signal_handler_disconnect(treeview, state->model_row_expanded);
-    if(state->model_row_collapsed)  g_signal_handler_disconnect(treeview, state->model_row_collapsed);
-  }
-
-  state->model = NULL;
-  state->model_row_inserted = 0;
-  state->model_row_deleted = 0;
-  state->model_row_changed = 0;
-  state->model_rows_reordered = 0;
-  state->model_row_expanded = 0;
-  state->model_row_collapsed = 0;
-}
-
-static void _widget_auto_disconnect_buffer(dt_gui_widget_auto_height_t *state)
-{
-  if(IS_NULL_PTR(state)) return;
-
-  GtkTextBuffer *buffer = state->buffer;
-  if(buffer)
-  {
-    if(state->buffer_changed) g_signal_handler_disconnect(buffer, state->buffer_changed);
-    g_object_remove_weak_pointer(G_OBJECT(buffer), (gpointer *)&state->buffer);
-  }
-
-  state->buffer = NULL;
-  state->buffer_changed = 0;
-}
-
-/**
- * @brief Window-height ceiling shared by the auto-size rule and the drag handle.
- *
- * @details The full main-window height: a resizable area may grow as tall as the window (the
- * parent panel scrolls to reach it). Content shorter than this still shrinks to fit, so this only
- * bounds how far the user can drag.
- */
-static gint _resizable_scroll_max_height(void)
-{
-  GtkWidget *win = dt_ui_main_window(darktable.gui->ui);
-  return win ? gtk_widget_get_allocated_height(win) : DT_PIXEL_APPLY_DPI(1000);
-}
-
-/**
- * @brief The single sizing rule for every dt_ui_scroll_wrap area.
- *
- * @details Height = clamp(min(content, cap), min_size, 75% window), where the cap is the user's
- * persisted height when set, otherwise the window ceiling so the area auto-grows to its content.
- * This makes small content shrink to fit, lets nested lists grow and have their parent panel scroll
- * until the user drags the handle to cap them, and snaps lists/textviews to whole rows to avoid
- * clipped half-rows. The computed bare (pre-padding) height is cached for the drag handle.
- */
-static void _resizable_scroll_apply(GtkWidget *w)
-{
-  dt_gui_widget_auto_height_t *state = g_object_get_data(G_OBJECT(w), DT_GUI_WIDGET_AUTO_HEIGHT_KEY);
-  if(IS_NULL_PTR(state)) return;
-
-  GtkWidget *sw = _search_parent_scrolled_window(w);
-  if(!GTK_IS_SCROLLED_WINDOW(sw)) return;
-
-  const gint max_height = _resizable_scroll_max_height();
-  const gint min_size = MAX(1, state->min_size);
-  const gboolean has_conf = state->config_str && dt_conf_key_exists(state->config_str);
-
-  gint height;
-  gint increment = 0;
-
-  if(state->mode == DT_UI_RESIZE_STATIC)
-  {
-    // Fixed height: the user's persisted size, or min_size as the default. Independent of content,
-    // so a content refresh (e.g. on hovering another thumbnail) never shifts the layout.
-    height = CLAMP(has_conf ? dt_conf_get_int(state->config_str) : min_size, min_size, max_height);
-  }
-  else
-  {
-    // Dynamic: fit to content, capped by the user's persisted height (or the window ceiling).
-    const gboolean row_based = GTK_IS_TREE_VIEW(w) || GTK_IS_TEXT_VIEW(w);
-    increment = row_based ? _get_container_row_heigth(w) : 0;
-
-    gint content = 0;
-    if(GTK_IS_TREE_VIEW(w))
-    {
-      const int rows = _treeview_count_visible_rows(GTK_TREE_VIEW(w), gtk_tree_view_get_model(GTK_TREE_VIEW(w)), NULL);
-      content = MAX(1, rows) * increment;
-    }
-    else if(GTK_IS_TEXT_VIEW(w))
-    {
-      const int rows = _textview_count_visible_rows(w);
-      content = MAX(1, rows) * increment;
-    }
-    else
-    {
-      gtk_widget_get_preferred_height(w, NULL, &content);
-    }
-
-    const gint cap = has_conf ? CLAMP(dt_conf_get_int(state->config_str), min_size, max_height) : max_height;
-    height = CLAMP(MIN(content, cap), min_size, max_height);
-
-    // snap to whole rows for lists/textviews to avoid clipped half-rows
-    if(increment > 0)
-    {
-      height += increment - 1;
-      height -= height % increment;
-    }
-  }
-  state->last_height = height;
-
-  GtkBorder padding;
-  gtk_style_context_get_padding(gtk_widget_get_style_context(sw),
-                                gtk_widget_get_state_flags(sw), &padding);
-
-  gint old_height = 0;
-  gtk_widget_get_size_request(sw, NULL, &old_height);
-  const gint new_height = height + padding.top + padding.bottom + (GTK_IS_TEXT_VIEW(w) ? 2 : 0);
-  if(new_height != old_height)
-    gtk_widget_set_size_request(sw, -1, new_height);
-}
-
-static void _widget_auto_update(GtkWidget *widget)
-{
-  _resizable_scroll_apply(widget);
-}
-
-static gboolean _resizable_scroll_draw(GtkWidget *w, cairo_t *cr, gpointer user_data)
-{
-  _resizable_scroll_apply(w);
-  return FALSE;
-}
-
-static void _resizable_scroll_realize(GtkWidget *w, gpointer user_data)
-{
-  _resizable_scroll_apply(w);
-}
-
-// Drag handle accessors: bare (pre-padding) target height, kept consistent with the sizing rule.
-static int _resizable_scroll_handle_get_size(gpointer user_data)
-{
-  GtkWidget *w = GTK_WIDGET(user_data);
-  const dt_gui_widget_auto_height_t *state = g_object_get_data(G_OBJECT(w), DT_GUI_WIDGET_AUTO_HEIGHT_KEY);
-  return state ? state->last_height : 0;
-}
-
-static int _resizable_scroll_handle_resize(int requested_size, gboolean finished, gpointer user_data)
-{
-  GtkWidget *w = GTK_WIDGET(user_data);
-  dt_gui_widget_auto_height_t *state = g_object_get_data(G_OBJECT(w), DT_GUI_WIDGET_AUTO_HEIGHT_KEY);
-  if(IS_NULL_PTR(state) || IS_NULL_PTR(state->config_str)) return requested_size;
-
-  const gint value = CLAMP(requested_size, MAX(1, state->min_size), _resizable_scroll_max_height());
-  dt_conf_set_int(state->config_str, value);
-  _resizable_scroll_apply(w);
-  return state->last_height;
-}
-
-static void _widget_auto_model_row_inserted(GtkTreeModel *model, GtkTreePath *path, GtkTreeIter *iter,
-                                              gpointer user_data)
-{
-  (void)model;
-  (void)path;
-  (void)iter;
-  _widget_auto_update(GTK_WIDGET(user_data));
-}
-
-static void _widget_auto_model_row_deleted(GtkTreeModel *model, GtkTreePath *path, gpointer user_data)
-{
-  (void)model;
-  (void)path;
-  _widget_auto_update(GTK_WIDGET(user_data));
-}
-
-static void _widget_auto_model_row_changed(GtkTreeModel *model, GtkTreePath *path, GtkTreeIter *iter,
-                                             gpointer user_data)
-{
-  (void)model;
-  (void)path;
-  (void)iter;
-  _widget_auto_update(GTK_WIDGET(user_data));
-}
-
-static void _widget_auto_model_rows_reordered(GtkTreeModel *model, GtkTreePath *path, GtkTreeIter *iter,
-                                                gint *new_order, gpointer user_data)
-{
-  (void)model;
-  (void)path;
-  (void)iter;
-  (void)new_order;
-  _widget_auto_update(GTK_WIDGET(user_data));
-}
-
-static void _widget_auto_model_row_expanded(GtkTreeView *tree_view, GtkTreeIter *expanded_iter, GtkTreePath *path,
-                                              gpointer user_data)
-{
-  (void)tree_view;
-  (void)expanded_iter;
-  (void)path;
-  _widget_auto_update(GTK_WIDGET(user_data));
-}
-
-static void _widget_auto_model_row_collapsed(GtkTreeView *tree_view, GtkTreeIter *collapsed_iter, GtkTreePath *path,
-                                               gpointer user_data)
-{
-  (void)tree_view;
-  (void)collapsed_iter;
-  (void)path;
-
-  // Recalculate tree view height after loading the data
-  _widget_auto_update(GTK_WIDGET(user_data));
-}
-
-static void _widget_auto_text_buffer_changed(GtkTextBuffer *buffer, gpointer user_data)
-{
-  (void)buffer;
-  _widget_auto_update(GTK_WIDGET(user_data));
-}
-
-static void _widget_auto_connect_model(GtkWidget *treeview)
-{
-  if(!GTK_IS_TREE_VIEW(treeview)) return;
-
-  dt_gui_widget_auto_height_t *state = g_object_get_data(G_OBJECT(treeview), DT_GUI_WIDGET_AUTO_HEIGHT_KEY);
-  if(IS_NULL_PTR(state)) return;
-
-  GtkTreeModel *model = gtk_tree_view_get_model(GTK_TREE_VIEW(treeview));
-  if(model == state->model) return;
-
-  _widget_auto_disconnect_model(state, treeview);
-  if(!GTK_IS_TREE_MODEL(model)) return;
-
-  state->model = model;
-  g_object_add_weak_pointer(G_OBJECT(model), (gpointer *)&state->model);
-  state->model_row_inserted =   g_signal_connect(model, "row-inserted",
-                                              G_CALLBACK(_widget_auto_model_row_inserted), treeview);
-  state->model_row_deleted =    g_signal_connect(model, "row-deleted",
-                                              G_CALLBACK(_widget_auto_model_row_deleted), treeview);
-  state->model_row_changed =    g_signal_connect(model, "row-changed",
-                                              G_CALLBACK(_widget_auto_model_row_changed), treeview);
-  state->model_rows_reordered = g_signal_connect(model, "rows-reordered",
-                                              G_CALLBACK(_widget_auto_model_rows_reordered), treeview);
-  state->model_row_expanded =   g_signal_connect(treeview, "row-expanded",
-                                              G_CALLBACK(_widget_auto_model_row_expanded), treeview);
-  state->model_row_collapsed =  g_signal_connect(treeview, "row-collapsed",
-                                              G_CALLBACK(_widget_auto_model_row_collapsed), treeview);
-}
-
-static void _widget_auto_connect_buffer(GtkWidget *textview)
-{
-  if(!GTK_IS_TEXT_VIEW(textview)) return;
-
-  dt_gui_widget_auto_height_t *state = g_object_get_data(G_OBJECT(textview), DT_GUI_WIDGET_AUTO_HEIGHT_KEY);
-  if(IS_NULL_PTR(state)) return;
-
-  GtkTextBuffer *buffer = gtk_text_view_get_buffer(GTK_TEXT_VIEW(textview));
-  if(buffer == state->buffer) return;
-
-  _widget_auto_disconnect_buffer(state);
-  if(!GTK_IS_TEXT_BUFFER(buffer)) return;
-
-  state->buffer = buffer;
-  g_object_add_weak_pointer(G_OBJECT(buffer), (gpointer *)&state->buffer);
-  state->buffer_changed = g_signal_connect(buffer, "changed",
-                                           G_CALLBACK(_widget_auto_text_buffer_changed), textview);
-}
-
-static void _widget_auto_on_model_changed(GObject *treeview, GParamSpec *pspec, gpointer user_data)
-{
-  (void)pspec;
-  (void)user_data;
-  _widget_auto_connect_model(GTK_WIDGET(treeview));
-  _widget_auto_update(GTK_WIDGET(treeview));
-}
-
-static void _widget_auto_on_buffer_changed(GObject *textview, GParamSpec *pspec, gpointer user_data)
-{
-  (void)pspec;
-  (void)user_data;
-  _widget_auto_connect_buffer(GTK_WIDGET(textview));
-  _widget_auto_update(GTK_WIDGET(textview));
-}
-
-static void _widget_auto_height_free(gpointer data)
-{
-  dt_gui_widget_auto_height_t *state = (dt_gui_widget_auto_height_t *)data;
-  if(IS_NULL_PTR(state)) return;
-  _widget_auto_disconnect_model(state, NULL);
-  _widget_auto_disconnect_buffer(state);
-  g_free(state->config_str);
-  dt_free(state);
-}
-
-void dt_gui_textview_set_padding(GtkTextView *textview)
-{
-  if(!GTK_IS_TEXT_VIEW(textview)) return;
-
-  gtk_text_view_set_left_margin(textview, DT_PIXEL_APPLY_DPI(4));
-  gtk_text_view_set_right_margin(textview, DT_PIXEL_APPLY_DPI(4));
-  gtk_text_view_set_top_margin(textview, DT_PIXEL_APPLY_DPI(2));
-  gtk_text_view_set_bottom_margin(textview, DT_PIXEL_APPLY_DPI(2));
-}
-
-/**
- * @brief Wrap a scrollable content widget in a recessed, vertically resizable scrolled window.
- *
- * @details Returns an overlay wrapping the scrolled window, with a themed drag grip floating on its
- * bottom edge (the same grip primitive used by panels and the histogram scope). The grip takes no
- * layout space and is invisible until hovered. Sizing follows @p mode: DT_UI_RESIZE_DYNAMIC auto-fits
- * the content up to the user height, DT_UI_RESIZE_STATIC keeps a fixed height (see _resizable_scroll_apply).
- *
- * The returned widget is the wrapper overlay, not the scrolled window; callers needing the inner
- * scrolled window (e.g. to tweak its scroll policy) must use dt_ui_scroll_wrap_get_scrolled_window().
- *
- * @param w content widget (treeview, textview or any container)
- * @param min_size minimum height floor in device pixels (also the static default before the user drags)
- * @param config_str conf key persisting the user-chosen height (copied internally)
- * @param mode DT_UI_RESIZE_DYNAMIC (auto-fit) or DT_UI_RESIZE_STATIC (fixed height)
- */
-GtkWidget *dt_ui_scroll_wrap(GtkWidget *w, gint min_size, char *config_str, dt_ui_resize_mode_t mode)
-{
-  GtkWidget *sw = gtk_scrolled_window_new(NULL, NULL);
-  gtk_scrolled_window_set_policy(GTK_SCROLLED_WINDOW(sw), GTK_POLICY_NEVER, GTK_POLICY_AUTOMATIC);
-  if(GTK_IS_TREE_VIEW(w)) dt_gui_add_class(sw, "dt_recessed_scroll");
-  gtk_container_add(GTK_CONTAINER(sw), w);
-
-  // Per-widget sizing state, freed with w.
-  dt_gui_widget_auto_height_t *state = calloc(1, sizeof(*state));
-  state->config_str = g_strdup(config_str);
-  state->min_size = MAX(1, DT_PIXEL_APPLY_DPI(min_size));
-  state->mode = mode;
-  g_object_set_data_full(G_OBJECT(w), DT_GUI_WIDGET_AUTO_HEIGHT_KEY, state, _widget_auto_height_free);
-
-  if(mode == DT_UI_RESIZE_DYNAMIC)
-  {
-    // Sizing triggers. Lists/textviews recompute from their model/buffer change signals (cheap and
-    // exact) plus a one-shot on realize, when row heights become accurate -- we deliberately avoid a
-    // per-draw recount, which would re-walk a large model on every redraw. Generic content has no
-    // such signals, so it recomputes on draw using its (GTK-cached) preferred height.
-    if(GTK_IS_TREE_VIEW(w))
-    {
-      g_signal_connect(w, "notify::model", G_CALLBACK(_widget_auto_on_model_changed), NULL);
-      g_signal_connect(w, "realize", G_CALLBACK(_resizable_scroll_realize), NULL);
-    }
-    else if(GTK_IS_TEXT_VIEW(w))
-    {
-      g_signal_connect(w, "notify::buffer", G_CALLBACK(_widget_auto_on_buffer_changed), NULL);
-      g_signal_connect(w, "realize", G_CALLBACK(_resizable_scroll_realize), NULL);
-    }
-    else
-    {
-      g_signal_connect(G_OBJECT(w), "draw", G_CALLBACK(_resizable_scroll_draw), NULL);
-    }
-    _widget_auto_connect_model(w);
-    _widget_auto_connect_buffer(w);
-  }
-  else
-  {
-    // Static: height is content-independent, so we only size once after realization (and on drag).
-    g_signal_connect(w, "realize", G_CALLBACK(_resizable_scroll_realize), NULL);
-  }
-
-  // Drag grip floating on the scrolled window's bottom edge (overlay): it takes no layout space,
-  // so the wrapper leaves no margin-like gap and stays aligned with neighbouring widgets. The grip
-  // is centered on the bottom border via CSS and is invisible until hovered.
-  GtkWidget *handle = dtgtk_resize_handle_new(GTK_ORIENTATION_VERTICAL, FALSE,
-                                                   _("Drag to resize"),
-                                                   _resizable_scroll_handle_get_size,
-                                                   _resizable_scroll_handle_resize, w);
-
-  GtkWidget *over = gtk_overlay_new();
-  gtk_container_add(GTK_CONTAINER(over), sw);
-  gtk_overlay_add_overlay(GTK_OVERLAY(over), handle);
-
-  _widget_auto_update(w);
-  return over;
-}
-
-/**
- * @brief Return the inner scrolled window of a dt_ui_scroll_wrap() wrapper, or NULL.
- */
-GtkWidget *dt_ui_scroll_wrap_get_scrolled_window(GtkWidget *wrapper)
-{
-  if(!GTK_IS_CONTAINER(wrapper)) return NULL;
-  GList *children = gtk_container_get_children(GTK_CONTAINER(wrapper));
-  GtkWidget *sw = NULL;
-  for(GList *l = children; l; l = g_list_next(l))
-  {
-    if(GTK_IS_SCROLLED_WINDOW(l->data))
-    {
-      sw = GTK_WIDGET(l->data);
-      break;
-    }
-  }
-  g_list_free(children);
-  return sw;
-}
-
-// ---- Resizable drawing area (the histogram-scope paradigm, made reusable) -------------------
-// A fixed-pixel-height GtkDrawingArea (or any widget sized by height-request) made vertically
-// resizable by the shared grip primitive, with the height persisted to config. Unlike the
-// scroll-wrap helper above, the content draws itself and is not scrolled; we only manage its
-// height-request, so the drawing code keeps reading the live allocation as usual.
-
-static const char *const DT_UI_RESIZABLE_AREA_KEY = "dt-ui-resizable-area";
-
-typedef struct dt_ui_resizable_area_t
-{
-  char *config_str;   // conf key persisting the user-chosen height (px); owned
-  int min_height;     // minimum height floor, device pixels
-  int last_height;    // last applied height, shared with the drag handle
-} dt_ui_resizable_area_t;
-
-static void _resizable_area_free(gpointer data)
-{
-  dt_ui_resizable_area_t *state = (dt_ui_resizable_area_t *)data;
-  if(IS_NULL_PTR(state)) return;
-  g_free(state->config_str);
-  dt_free(state);
-}
-
-static int _resizable_area_get_size(gpointer user_data)
-{
-  GtkWidget *area = GTK_WIDGET(user_data);
-  const dt_ui_resizable_area_t *state = g_object_get_data(G_OBJECT(area), DT_UI_RESIZABLE_AREA_KEY);
-  if(state) return state->last_height;
-  return gtk_widget_get_allocated_height(area);
-}
-
-static int _resizable_area_resize(int requested_size, gboolean finished, gpointer user_data)
-{
-  GtkWidget *area = GTK_WIDGET(user_data);
-  dt_ui_resizable_area_t *state = g_object_get_data(G_OBJECT(area), DT_UI_RESIZABLE_AREA_KEY);
-  if(IS_NULL_PTR(state)) return requested_size;
-
-  const int height = CLAMP(requested_size, MAX(1, state->min_height), _resizable_scroll_max_height());
-  state->last_height = height;
-  gtk_widget_set_size_request(area, -1, height);
-  if(finished && state->config_str) dt_conf_set_int(state->config_str, height);
-  return height;
-}
-
-GtkWidget *dt_ui_resizable_drawing_area(GtkWidget *area, char *config_str, int default_height, int min_height)
-{
-  dt_ui_resizable_area_t *state = calloc(1, sizeof(*state));
-  state->config_str = g_strdup(config_str);
-  state->min_height = MAX(1, DT_PIXEL_APPLY_DPI(min_height));
-
-  int height = (config_str && dt_conf_key_exists(config_str)) ? dt_conf_get_int(config_str)
-                                                              : DT_PIXEL_APPLY_DPI(default_height);
-  height = CLAMP(height, state->min_height, _resizable_scroll_max_height());
-  state->last_height = height;
-  g_object_set_data_full(G_OBJECT(area), DT_UI_RESIZABLE_AREA_KEY, state, _resizable_area_free);
-  gtk_widget_set_size_request(area, -1, height);
-
-  // Drag grip floating on the area's bottom edge (an overlay, not a packed sibling), so it takes
-  // no layout space -- the area stays flush with neighbouring widgets, no margin-like gap. It sits
-  // over the graph's bottom inset/axis margin (graphs reserve one), invisible until hovered.
-  GtkWidget *handle = dtgtk_resize_handle_new(GTK_ORIENTATION_VERTICAL, FALSE,
-                                                   _("Drag to resize"),
-                                                   _resizable_area_get_size, _resizable_area_resize, area);
-
-  GtkWidget *over = gtk_overlay_new();
-  gtk_container_add(GTK_CONTAINER(over), area);
-  gtk_overlay_add_overlay(GTK_OVERLAY(over), handle);
-  return over;
-}
-
-gboolean dt_gui_container_has_children(GtkContainer *container)
-{
-  g_return_val_if_fail(GTK_IS_CONTAINER(container), FALSE);
-  GList *children = gtk_container_get_children(container);
-  gboolean has_children = !IS_NULL_PTR(children);
-  g_list_free(children);
-  children = NULL;
-  return has_children;
-}
-
-int dt_gui_container_num_children(GtkContainer *container)
-{
-  g_return_val_if_fail(GTK_IS_CONTAINER(container), FALSE);
-  GList *children = gtk_container_get_children(container);
-  int num_children = g_list_length(children);
-  g_list_free(children);
-  children = NULL;
-  return num_children;
-}
-
-GtkWidget *dt_gui_container_first_child(GtkContainer *container)
-{
-  g_return_val_if_fail(GTK_IS_CONTAINER(container), NULL);
-  GList *children = gtk_container_get_children(container);
-  GtkWidget *child = children ? (GtkWidget*)children->data : NULL;
-  g_list_free(children);
-  children = NULL;
-  return child;
-}
-
-GtkWidget *dt_gui_container_nth_child(GtkContainer *container, int which)
-{
-  g_return_val_if_fail(GTK_IS_CONTAINER(container), NULL);
-  GList *children = gtk_container_get_children(container);
-  GtkWidget *child = (GtkWidget*)g_list_nth_data(children, which);
-  g_list_free(children);
-  children = NULL;
-  return child;
-}
-
-static void _remove_child(GtkWidget *widget, gpointer data)
-{
-  gtk_container_remove((GtkContainer*)data, widget);
-}
-
-void dt_gui_container_remove_children(GtkContainer *container)
-{
-  g_return_if_fail(GTK_IS_CONTAINER(container));
-  gtk_container_foreach(container, _remove_child, container);
-}
-
-static void _delete_child(GtkWidget *widget, gpointer data)
-{
-  (void)data;  // avoid unreferenced-parameter warning
-  gtk_widget_destroy(widget);
-}
-
-void dt_gui_container_destroy_children(GtkContainer *container)
-{
-  g_return_if_fail(GTK_IS_CONTAINER(container));
-  gtk_container_foreach(container, _delete_child, NULL);
-}
-
-GtkWidget *dt_gui_get_popup_relative_widget(GtkWidget *widget, GdkRectangle *rect)
-{
-  if(IS_NULL_PTR(widget)) return NULL;
-
-  GtkWidget *relative = widget;
-
-  // Wayland only accepts the top-most enclosing popup as transient parent.
-  for(GtkWidget *parent = gtk_widget_get_parent(widget); parent; parent = gtk_widget_get_parent(parent))
-    if(GTK_IS_POPOVER(parent)) relative = parent;
-
-  if(rect)
-  {
-    rect->x = 0;
-    rect->y = 0;
-    rect->width = MAX(gtk_widget_get_allocated_width(widget), 1);
-    rect->height = MAX(gtk_widget_get_allocated_height(widget), 1);
-
-    if(relative != widget
-       && !gtk_widget_translate_coordinates(widget, relative, 0, 0, &rect->x, &rect->y))
-    {
-      rect->x = 0;
-      rect->y = 0;
-    }
-  }
-
-  return relative;
-}
-
-void dt_gui_menu_popup(GtkMenu *menu, GtkWidget *button, GdkGravity widget_anchor, GdkGravity menu_anchor)
-{
-  gtk_widget_show_all(GTK_WIDGET(menu));
-
-  GdkEvent *event = gtk_get_current_event();
-  if(button)
-  {
-    GdkRectangle rect = { 0 };
-    GtkWidget *relative = dt_gui_get_popup_relative_widget(button, &rect);
-
-    if(relative && relative != button && gtk_widget_get_window(relative))
-      gtk_menu_popup_at_rect(menu, gtk_widget_get_window(relative), &rect, widget_anchor, menu_anchor, event);
-    else
-      gtk_menu_popup_at_widget(menu, button, widget_anchor, menu_anchor, event);
-  }
-  else
-  {
-    if(IS_NULL_PTR(event))
-    {
-      event = gdk_event_new(GDK_BUTTON_PRESS);
-      event->button.device = gdk_seat_get_pointer(gdk_display_get_default_seat(gdk_display_get_default()));
-      event->button.window = gtk_widget_get_window(GTK_WIDGET(darktable.gui->ui->main_window));
-      g_object_ref(event->button.window);
-    }
-
-    gtk_menu_popup_at_pointer(menu, event);
-  }
-  gdk_event_free(event);
-}
-
-static void _popover_set_relative_to_topmost_parent(GtkPopover *popover, GtkWidget *button)
-{
-  GdkRectangle rect = { 0 };
-  GtkWidget *relative = dt_gui_get_popup_relative_widget(button, &rect);
-  gtk_popover_set_relative_to(popover, relative ? relative : button);
-  gtk_popover_set_pointing_to(popover, &rect);
-}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 // draw rounded rectangle
 
-gboolean dt_gui_search_start(GtkWidget *widget, GdkEventKey *event, GtkSearchEntry *entry)
-{
-  if(gtk_search_entry_handle_event(entry, (GdkEvent *)event))
-  {
-    gtk_entry_grab_focus_without_selecting(GTK_ENTRY(entry));
-    return TRUE;
-  }
-
-  return FALSE;
-}
-
-void dt_gui_search_stop(GtkSearchEntry *entry, GtkWidget *widget)
-{
-  gtk_widget_grab_focus(widget);
-
-  gtk_entry_set_text(GTK_ENTRY(entry), "");
-
-  if(GTK_IS_TREE_VIEW(widget))
-  {
-    GtkTreePath *path = NULL;
-    gtk_tree_view_get_cursor(GTK_TREE_VIEW(widget), &path, NULL);
-    gtk_tree_selection_select_path(gtk_tree_view_get_selection(GTK_TREE_VIEW(widget)), path);
-    gtk_tree_path_free(path);
-  }
-}
-
-static void _collapsible_set_states(dt_gui_collapsible_section_t *cs, gboolean active)
-{
-  if(active)
-  {
-    // We don't apply the GTK_STATE_SELECTED flag to the container here because it would
-    // be inherited by all children, which would mess up the state of checkboxes and togglebuttons.
-    dt_gui_add_class(GTK_WIDGET(cs->expander), "active");
-  }
-  else
-  {
-    gtk_widget_set_state_flags(GTK_WIDGET(cs->expander), GTK_STATE_FLAG_NORMAL, TRUE);
-    dt_gui_remove_class(GTK_WIDGET(cs->expander), "active");
-  }
-}
-
-static void _collapsible_container_show(GtkWidget *widget, gpointer user_data)
-{
-  /* Called whenever the container receives a "show" event, including from gtk_widget_show_all().
-   * If the toggle is not active the section should remain collapsed, so we re-hide the container
-   * immediately. By the time this fires, show_all has already recursed into the children and
-   * set their visible flags, so a later expand will find all children ready to display. */
-  dt_gui_collapsible_section_t *cs = (dt_gui_collapsible_section_t *)user_data;
-  if(!gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(cs->toggle)))
-    gtk_widget_hide(widget);
-}
-
-static void _coeffs_button_changed(GtkDarktableToggleButton *widget, gpointer user_data)
-{
-  dt_gui_collapsible_section_t *cs = (dt_gui_collapsible_section_t *)user_data;
-
-  const gboolean active = gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(cs->toggle));
-  dtgtk_expander_set_expanded(DTGTK_EXPANDER(cs->expander), active);
-  dtgtk_togglebutton_set_paint(DTGTK_TOGGLEBUTTON(cs->toggle), dtgtk_cairo_paint_solid_arrow,
-                               (active ? CPF_DIRECTION_DOWN : CPF_DIRECTION_LEFT), NULL);
-  dt_conf_set_bool(cs->confname, active);
-  _collapsible_set_states(cs, active);
-}
-
-static void _coeffs_expander_click(GtkWidget *widget, GdkEventButton *e, gpointer user_data)
-{
-  if(e->type == GDK_2BUTTON_PRESS || e->type == GDK_3BUTTON_PRESS) return;
-
-  dt_gui_collapsible_section_t *cs = (dt_gui_collapsible_section_t *)user_data;
-
-  const gboolean active = gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(cs->toggle));
-  gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(cs->toggle), !active);
-  _collapsible_set_states(cs, !active);
-}
-
-void dt_gui_update_collapsible_section(dt_gui_collapsible_section_t *cs)
-{
-  const gboolean active = gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(cs->toggle));
-  dtgtk_togglebutton_set_paint(DTGTK_TOGGLEBUTTON(cs->toggle), dtgtk_cairo_paint_solid_arrow,
-                               (active ? CPF_DIRECTION_DOWN : CPF_DIRECTION_LEFT), NULL);
-  dtgtk_expander_set_expanded(DTGTK_EXPANDER(cs->expander), active);
-
-  if(active)
-    gtk_widget_show(GTK_WIDGET(cs->container));
-  else
-    gtk_widget_hide(GTK_WIDGET(cs->container));
-
-  _collapsible_set_states(cs, active);
-}
-
-void dt_gui_hide_collapsible_section(dt_gui_collapsible_section_t *cs)
-{
-  gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(cs->toggle), FALSE);
-  gtk_widget_hide(GTK_WIDGET(cs->container));
-  _collapsible_set_states(cs, FALSE);
-}
-
-void dt_gui_new_collapsible_section(dt_gui_collapsible_section_t *cs,
-                                    const char *confname, const char *label,
-                                    GtkBox *parent, GtkPackType pack)
-{
-  const gboolean expanded = dt_conf_get_bool(confname);
-
-  cs->confname = confname;
-  cs->parent = parent;
-
-  // collapsible section header
-  GtkWidget *destdisp_head = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, DT_GUI_BOX_SPACING);
-  GtkWidget *header_evb = gtk_event_box_new();
-  cs->label = dt_ui_section_label_new(label);
-  dt_gui_add_class(destdisp_head, "dt_section_expander");
-  gtk_container_add(GTK_CONTAINER(header_evb), cs->label);
-
-  cs->toggle = dtgtk_togglebutton_new(dtgtk_cairo_paint_solid_arrow,
-                                      (expanded ? CPF_DIRECTION_DOWN : CPF_DIRECTION_LEFT), NULL);
-  gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(cs->toggle), expanded);
-  dt_gui_add_class(cs->toggle, "dt_ignore_fg_state");
-
-  cs->container = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING));
-  gtk_widget_set_name(GTK_WIDGET(cs->container), "collapsible");
-  gtk_box_pack_start(GTK_BOX(destdisp_head), header_evb, TRUE, TRUE, 0);
-  gtk_box_pack_start(GTK_BOX(destdisp_head), cs->toggle, FALSE, FALSE, 0);
-
-  cs->expander = dtgtk_expander_new(destdisp_head, GTK_WIDGET(cs->container));
-  /* gtk_widget_show_all() on the parent (called by lib modules after gui_init) recurses into
-   * the container and would override the collapsed state set by dtgtk_expander_set_expanded.
-   * Connect to "show" so we can re-apply the correct visibility right after show_all touches it. */
-  g_signal_connect(G_OBJECT(cs->container), "show",
-                   G_CALLBACK(_collapsible_container_show), (gpointer)cs);
-  // Pack at the requested side so callers control ordering at insertion time.
-  if(pack == GTK_PACK_START)
-    gtk_box_pack_start(GTK_BOX(cs->parent), cs->expander, FALSE, FALSE, 0);
-  else
-    gtk_box_pack_end(GTK_BOX(cs->parent), cs->expander, FALSE, FALSE, 0);
-  dtgtk_expander_set_expanded(DTGTK_EXPANDER(cs->expander), expanded);
-  gtk_widget_set_name(cs->expander, "collapse-block");
-
-  g_signal_connect(G_OBJECT(cs->toggle), "toggled",
-                   G_CALLBACK(_coeffs_button_changed),  (gpointer)cs);
-
-  g_signal_connect(G_OBJECT(header_evb), "button-release-event",
-                   G_CALLBACK(_coeffs_expander_click),
-                   (gpointer)cs);
-}
 
 
-GtkBox * attach_popover(GtkWidget *widget, const char *icon, GtkWidget *content)
-{
-  // Create the wrapping box and add the original widget to it
-  GtkWidget *box = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, DT_GUI_BOX_SPACING);
-  gtk_box_pack_start(GTK_BOX(box), widget, FALSE, FALSE, 0);
 
-  // Create the info icon button that will trigger the popover
-  GtkWidget *button = gtk_menu_button_new();
-  dt_gui_add_class(button, "popover-button");
-  GtkWidget *image = gtk_image_new_from_icon_name(icon, GTK_ICON_SIZE_BUTTON);
-  gtk_button_set_image(GTK_BUTTON(button), image);
-  gtk_widget_set_hexpand(button, FALSE);
-  gtk_widget_set_vexpand(button, FALSE);
-  gtk_widget_set_size_request(button, DT_PIXEL_APPLY_DPI(16), DT_PIXEL_APPLY_DPI(16));
-  gtk_box_pack_start(GTK_BOX(box), button, FALSE, FALSE, 0);
 
-  // Create the content of the popover
-  GtkWidget *popover_box = gtk_box_new(GTK_ORIENTATION_VERTICAL, DT_GUI_BOX_SPACING);
-  gtk_box_pack_start(GTK_BOX(popover_box), content, FALSE, FALSE, 0);
 
-  // Wrap the content into a popover and attach it to the button
-  GtkWidget *popover = gtk_popover_new(button);
-  gtk_container_add(GTK_CONTAINER(popover), popover_box);
-  gtk_popover_set_modal(GTK_POPOVER(popover), FALSE);
-  g_signal_connect(G_OBJECT(popover), "show", G_CALLBACK(_popover_set_relative_to_topmost_parent), button);
-  gtk_menu_button_set_popover(GTK_MENU_BUTTON(button), popover);
-  gtk_widget_show_all(popover_box);
 
-  return GTK_BOX(box);
-}
 
-GtkBox * attach_help_popover(GtkWidget *widget, const char *label)
-{
-  // Create the content of the popover
-  GtkWidget *popover_label = gtk_label_new(label);
-  gtk_label_set_line_wrap(GTK_LABEL(popover_label), TRUE);
-  gtk_label_set_max_width_chars(GTK_LABEL(popover_label), 60);
-  return attach_popover(widget, "help-about", popover_label);
-}
 
-static gboolean _text_entry_focus_in_event(GtkWidget *self, GdkEventFocus event, gpointer user_data)
-{
-  dt_accels_disable(darktable.gui->accels, TRUE);
-  return FALSE;
-}
 
-static gboolean _text_entry_focus_out_event(GtkWidget *self, GdkEventFocus event, gpointer user_data)
-{
-  dt_accels_disable(darktable.gui->accels, FALSE);
-  return FALSE;
-}
 
-static gboolean _text_entry_key_pressed(GtkWidget *widget, GdkEventKey *event, gpointer user_data)
-{
-  if(event->keyval == GDK_KEY_Escape)
-  {
-    dt_gui_refocus_center();
-    return TRUE;
-  }
-  return FALSE;
-}
 
-void dt_accels_disconnect_on_text_input(GtkWidget *widget)
-{
-  gtk_widget_add_events(widget, GDK_FOCUS_CHANGE_MASK);
-  g_signal_connect(G_OBJECT(widget), "focus-in-event", G_CALLBACK(_text_entry_focus_in_event), NULL);
-  g_signal_connect(G_OBJECT(widget), "focus-out-event", G_CALLBACK(_text_entry_focus_out_event), NULL);
-  g_signal_connect(G_OBJECT(widget), "key-press-event", G_CALLBACK(_text_entry_key_pressed), NULL);
-}
+
+
+
+
 
 
 void dt_gui_refocus_center()
