+++ src/widgets/accelerators.h
@@ -391,6 +391,15 @@ void dt_accels_search(dt_accels_t *accels, GtkWindow *main_window, GtkWidget *an
 void dt_accels_set_global(dt_accels_t *accels);
 dt_accels_t *dt_accels_get_global(void);
 
+/**
+ * @brief Disconnect accels while a text or search entry has the focus, and reconnect them when
+ * it loses it. This is what makes one-key shortcuts safe to have at all. Escape hands focus
+ * back to the host's main area via dt_widget_refocus().
+ *
+ * @param widget the entry to guard.
+ */
+void dt_accels_disconnect_on_text_input(GtkWidget *widget);
+
 /** Vertical offset, in pixels, below which a shortcut window should be placed (the host's
  *  top panel height). Unregistered: 0. */
 typedef gint (*dt_accels_top_offset_handler_t)(void);
